refactor: separate client-side fetch and server-side SDK logic
